Prototype CH1L6, CH2L3, Hard SECZ
Found an old backup from the pwad days of CH1L6, so I duct-taped it on to the current portion. Chapter One is now completely playable start-to-finish. CH2L3 ends with a text intermission, but then the engine quits to console because I haven't made CH2L4 yet. Zombie Security Guards also can throw grenades on Hard or Impossible skills.